feat: convert spinner to async API
So the reason we have a separate spinner process in the first place is because broccoli's build API often results in blocking Node's single thread. So to maintain any kind of UI responsiveness (i.e. show a CLI spinner), we need to separate the spinner and the build into different Node processes. The spinner is much easier to separate given it's relative simplicity, so that was the easier path.

Unfortunately, moving to a child process introduces asynchrony (well, that's kinda the point ...). So spinner methods are now async, and the parent process will only resolve those async methods once the child process sends an ack, ensuring operations are applied in order.