feat: significantly improve CLI startup times
Move a few heavier modules from static imports to lazy `require()`s.
Also remove support for globally installed addons via yarn *and* npm.
If Yarn is detected, Denali will search Yarn for global addons, otherwise
it wills search npm, but not both. This lets Yarn users avoid the
delays that npm forces (because we need to shell out for `npm root`, which
eats 500ms).

We could make this behavior configurable in the future, once we have
a ~/.denali config file