fix: delay addon discovery until we know it will be needed (i.e. the addon will be built)
This allows us to always load devDeps as addons. We used to only load devDeps for root builders, but
that becomes problematic for addons that are build on the fly, i.e. symlinked or git dep addons. SO
now we always load devDeps for an addon that is being built, since if the addon its being built, it
will likely need any devDep addons like language compilers