fix: copy top level files into dummy throwaway app for command acceptance tests
Previously, we only copied the dist folder into the throwaway app's copy of the addon.
This is problematic because:

1. without a package.json, the addon won't get picked up as a node_module, and
2. if the addon has a processParent hook in it's denali-build.js file, it
   won't get picked up, and the throwaway app won't compile correctly

This changes the logic to copy all the standard publishable parts into
the throwaway app's copy of the addon, to better simulate a true installed
package.

This could become problematic in the future if an addon has some other
top level files it needs, but we'll cross that bridge then