fix: ignore errors when unable to communicate with spinner process
It's possible to try to send a message to the child spinner process after
it has died. Usually, this is the result of recieving a SIGINT during
the middle of a blocking operation (i.e. a broccoli build). The child
process will see the SIGINT immediately and exit, but the parent CLI
process only sees it once it finishes the blocking operation. Since
Node finishes current tick of the event loop before exiting, it attempts
to flush a message to the child process, which has already exited,
resulting in an error.